# swagger_server/models/track.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Track(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, idtrack: int=None, track: bytearray=None):  # noqa: E501
        """Track - a model defined in Swagger

        :param idtrack: The idtrack of this Track.  # noqa: E501
        :type idtrack: int
        :param track: The track of this Track.  # noqa: E501
        :type track: bytearray
        """
        self.swagger_types = {
            'idtrack': int,
            'track': bytearray
        }

        self.attribute_map = {
            'idtrack': 'idtrack',
            'track': 'track'
        }
        self._idtrack = idtrack
        self._track = track

    @classmethod
    def from_dict(cls, dikt) -> 'Track':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Track of this Track.  # noqa: E501
        :rtype: Track
        """
        return util.deserialize_model(dikt, cls)

    @property
    def idtrack(self) -> int:
        """Gets the idtrack of this Track.


        :return: The idtrack of this Track.
        :rtype: int
        """
        return self._idtrack

    @idtrack.setter
    def idtrack(self, idtrack: int):
        """Sets the idtrack of this Track.


        :param idtrack: The idtrack of this Track.
        :type idtrack: int
        """
        if idtrack is None:
            raise ValueError("Invalid value for `idtrack`, must not be `None`")  # noqa: E501

        self._idtrack = idtrack

    @property
    def track(self) -> bytearray:
        """Gets the track of this Track.


        :return: The track of this Track.
        :rtype: bytearray
        """
        return self._track

    @track.setter
    def track(self, track: bytearray):
        """Sets the track of this Track.


        :param track: The track of this Track.
        :type track: bytearray
        """
        if track is None:
            raise ValueError("Invalid value for `track`, must not be `None`")  # noqa: E501

        self._track = track